requirements.txt
python-dotenv
aiogram
asqlite3
aiopygismeteo